style.sh && merge master (#2023)
* Update versions for Release 5.11.0

* Update Firestore Generated Protos (#1972)

* Remove cruft from project file (#1975)

* Lock reads and writes to session map

* Changed FDL to use compatible __typeof__() operator. (#1982)

typeof() is not supported in some C language dialects which causes
issues when integrating FDL in projects that do use non-GNU compiler modes.

For example, this causes problems when building in some versions of Unity
https://github.com/firebase/quickstart-unity/issues/228

* Changed FDL to use compatible __typeof__() operator. (#1982)

typeof() is not supported in some C language dialects which causes
issues when integrating FDL in projects that do use non-GNU compiler modes.

For example, this causes problems when building in some versions of Unity
https://github.com/firebase/quickstart-unity/issues/228

* Bump DynamicLinks patch version

* Add FDL to release manifest

* Bump GoogleUtilities patch version to deliver fix for #1964 (#1987)

* Wrap diagnostics notification in collection flag check. (#1979)

* Wrap diagnostics notification in collection flag check.

Some of the diagnostics notifications were missed and not covered by
the data collection flag.

* Remove redundant notification call, move Core diagnostics API call.

* Removed configure with no options test.

* Queue Storage Tasks on separate queue (#1981)

* Increase timeout for testDispatchAfterDelay (#1988)

* Update Storage Changelog (#1989)

* Add missing FirebaseStorage release note (#1990)

* Allow semicolon in file name (#1991)

*  Remove unnecessary notification flag (#1993)

* Remove unnecessary notification flag.

This was added when the Google pod could configure Firebase but the
Google pod is deprecated and can only work with Firebase 3.X. These
flags and conditional checks can be safely removed.

* Resolve issues from commit split.

* Style fixes.

* Reduce singleton usage in FIRApp tests. (#1995)

* Reduce singleton usage in FIRApp tests.

There have been some issues while creating new tests of conflicts with
mocks of classes and instances, this should alleviate some of those
conflicts in the future.

* Remove bad style changes.

* Use default app name flag instead of local variable.

* Comply with c99 standard (#1992)

* Trigger travis for Firestore podspec changes

* Use C99-compatible __typeof__ instead of typeof (#1985)

`typeof` is only defined if you compile with GNU extensions, while
`__typeof__` is always available.

This is the Firestore equivalent of #1982.

Note that Firestore won't yet build in this mode because among other
things the Objective-C gRPC still uses `typeof`. Once we eliminate that
dependency this might become possible.

* Adding AppCode Diff (#1996)

* Remove warning (#1999)

* Add InAppMessaging to Carthage template (#2006)

* Restore SafariServices framework (#2002)

* SafariServices not available on tvOS and not used on osx

* Force Firestore to conform to C99 and C++11. (#2001)

Note that c++0x is how Xcode spells c++11.

Also fix an issue where we were accidentally using a C++14 feature.

* Changing the internal testing repo (#2003)

* Clean up test. The issue has already been fixed and we are now running integration test with actual server instead of hexa. (#2007)

* gRPC: replace Objective-C implementation with the new C++ implementation (#1968)

* add support for SSL disabled to `GrpcConnection` (unfortunately, there
  currently is no way to verify this change actually works);
* make gRPC calls using the C++ implementation:
* make `FSTRemoteStore` create C++ streams instead of their Objective-C
  counterparts;
* port https://github.com/firebase/firebase-js-sdk/pull/1041: streams
  are now never recreated, only restarted;
* make `FSTDatastore` delegate server calls to the C++ implementation;
* port `MockWatchStream` and `MockWriteStream` to C++ (`FSTMockDatastore` is
  still in place, because `Datastore` is not fully ported yet);
* no longer generate Objective-C gRPC service definitions from protos;
* remove all references to Objective-C gRPC client;
* check in gRPC root certificates file and load it at runtime (the check-in part
  is temporary until gRPC-C++.podspec provides the certificate). This makes SSL
  work.

* Add component system documentation.

* Fixed markdown layout issue.

* Remove trailing whitespaces.

* Add table of contents.

* Remove extra parentheses from TOC.

* Renamed SDKs to frameworks and products for accuracy.

* Updated the Carthage installation instructions (#2012)

* Add Rome instructions (#2014)

* Attempt to fix frequent Auth Unit Test flake on OSX (#2017)

* Increase timeouts in attempt to eliminate travis flakes (#2016)

* Don't rely on test always being in foreground (#2021)

* Fix log overflow in continuous fuzz testing (#2020)

Prevent generating too many "Unrecognized selector" console messages that eventually make Travis kill the job due to log exceeding limits.

* style.sh

+++ Firestore/Source/Local/FSTMemoryPersistence.mm
@@ -246,7 +246,7 @@ NS_ASSUME_NONNULL_BEGIN
   for (const auto &key : removed) {
     _sequenceNumbers.erase(key);
   }
-  return removed.size();
+  return static_cast<int>(removed.size());
 }
 
 - (void)addReference:(const DocumentKey &)key {
